backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-core

huggingface-hub

faiss-cpu
numpy

pydantic
python-dotenv
pypdf
requests

transformers

gTTS
soundfile